+++ src/test/java/domain/RoundTest.java
@@ -1,0 +1,40 @@
+package domain;
+
+import org.junit.Test;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+public class RoundTest {
+
+    @Test
+    public void should_add_new_Frame_when_appropriate_trying() {
+        Round round = new Round();
+        round.trying(1);
+        round.trying(5); // frame end
+        round.trying(6); // new frame add
+        round.trying(0); // frame end
+        round.trying(3); // new frame add
+        round.trying(7); // spare, frame end
+        round.trying(10); // new frame add, strike, frame end
+        assertThat(round.getRoundFrames().size(), is(5));
+    }
+
+    @Test
+    public void should_add_10th_frame_when_9th_frame_end() {
+        Round round = new Round();
+        round.trying(1); // game start 1st frame
+        round.trying(2); // frame end
+        round.trying(4); // 2nd frame add
+        round.trying(6); // frame end, spare
+        round.trying(10); // 3rd frame add, strike, frame end
+        round.trying(10); // 4th frame add, strike, frame end
+        round.trying(10); // 5th frame add, strike, frame end
+        round.trying(10); // 6th frame add, strike, frame end
+        round.trying(10); // 7th frame add, strike, frame end
+        round.trying(10); // 8th frame add, strike, frame end
+        round.trying(10); // 9th frame add, strike, frame end
+        assertThat(round.getRoundFrames().get(9), is(new LastFrame()));
+    }
+
+}
