+++ src/test/java/domain/ScoreTest.java
@@ -1,0 +1,4 @@
+package domain;
+
+public class ScoreTest {
+}
