+++ src/test/java/domain/ScoresTest.java
@@ -1,0 +1,7 @@
+package domain;
+
+public class ScoresTest {
+
+
+
+}
